非对称加密

+++ userservice/src/main/java/com/example/userservice/service/UserService.java
@@ -3,11 +3,13 @@ package com.example.userservice.service;
 import com.example.userservice.db.entity.UserEntity;
 import com.example.util.R;
 
+import java.util.ArrayList;
 import java.util.Map;
 
 public interface UserService {
 
     public String login(UserEntity userEntity);
 
-    public int register(Map param);
+    public long register(Map param);
+
 }
